+++ JamGame/Assets/ScriptableObjects/Level Configs/LevelConfig.asset
@@ -12,7 +12,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 51c39a70600ea274ebd8278a07991c85, type: 3}
   m_Name: LevelConfig
   m_EditorClassIdentifier: 
-  days: []
+  days:
+  - rid: 1908735195745353728
   defaultDay:
     dayActions:
     - rid: 1908735064151687181
@@ -95,3 +96,12 @@ MonoBehaviour:
       data:
         duration: 2
         text: Day End
+    - rid: 1908735195745353728
+      type: {class: DayConfig, ns: Level.Config, asm: Assembly-CSharp}
+      data:
+        dayActions:
+        - rid: 1908735195745353729
+    - rid: 1908735195745353729
+      type: {class: LoadLevel, ns: Level.Config, asm: Assembly-CSharp}
+      data:
+        buildingConfig: {fileID: 11400000, guid: 47361f59a2acd9e419a8ed3203c4cbc4, type: 2}
