+++ JamGame/Assets/Scripts/TileBuilder/Controller/ControllerImpl.cs
@@ -127,5 +127,15 @@ namespace TileBuilder.Controller
         {
             return tileBuilder.Bounds;
         }
+
+        public void LoadBuildingFromConfig(BuildingConfig buildingConfig)
+        {
+            _ = tileBuilder.ExecuteCommand(new RemoveAllRooms());
+            foreach (TileConfig tileConfig in buildingConfig.TilePlaceConfigs)
+            {
+                CoreModel core = CoreModel.InstantiateCoreModel(tileConfig);
+                _ = tileBuilder.ExecuteCommand(new DropRoom(core));
+            }
+        }
     }
 }
